Dragon Quest 6 snap improvement
Changing from fairly generic existing snap (which is from a few moments into the game) with the highly distinct opening shot of the game which is also one of the coziest scenes in RPG history.  It also shows the unusual light-glow effect that the artists put on the tree leaf tiles.